Migration to MariaDB

+++ db/migrate/20201109133817_modify_activities_column_name.rb
@@ -1,6 +1,6 @@
 class ModifyActivitiesColumnName < ActiveRecord::Migration[6.0]
   def change
-    add_column :activities, :user_id, :integer
+    add_column :activities, :user_id, :bigint
     remove_column :activities, :user
   end
 end
